+++ crates/hir/src/lib.rs
@@ -32,9 +32,9 @@ mod item;
 mod name_resolver;
 pub mod testing;
 
-use std::{collections::HashMap, marker::PhantomData};
+use std::collections::HashMap;
 
-use ast::AstNode;
+use ast::{AstNode, AstPtr};
 pub use body::{BindingId, BodyLower, ExprId, FunctionBodyId, HirFileDatabase};
 pub use db::{HirMasterDatabase, Jar};
 pub use input::{FixtureDatabase, NailFile, SourceDatabase, SourceDatabaseTrait};
@@ -378,36 +378,6 @@ pub fn build_green_node(db: &dyn HirMasterDatabase, nail_file: NailFile) -> Nail
     }
 
     NailGreenNode::new(db, nail_file, parse_result.green_node)
-}
-
-/// ASTノードのIDです。
-///
-/// 1ファイル内でユニークです。
-/// IDからASTを参照し、ファイル内における位置を取得することができます。
-#[derive(Debug, PartialEq, Eq, Hash)]
-pub struct AstPtr<T: AstNode> {
-    /// ASTノード
-    pub node: ast::SyntaxNodePtr,
-    _ty: PhantomData<T>,
-}
-impl<T: AstNode> Clone for AstPtr<T> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-impl<T: AstNode> Copy for AstPtr<T> {}
-/// `AstPtr`はスレッドセーフです。
-/// `AstNode`はスレッドセーフではありませんが、`AstPtr`は`AstNode`の型情報のみを参照するだけのためスレッドセーフです。
-unsafe impl<T: AstNode> Send for AstPtr<T> {}
-unsafe impl<T: AstNode> Sync for AstPtr<T> {}
-impl<T: AstNode> AstPtr<T> {
-    /// 新しいASTポインタを作成します。
-    pub fn new(node: &T) -> Self {
-        Self {
-            node: ast::SyntaxNodePtr::new(node.syntax()),
-            _ty: PhantomData,
-        }
-    }
 }
 
 /// 式のAST位置です。
